Instrument the import command
This command is short lived, and @KtorZ mentioned it might go away;
still, to experiment with structured spans that go a bit deeper than
elsewhere in the app, I went ahead and matched the natural structure of
the code that was present in the comments / serialization format.

Using {} blocks allows us to avoid calling .exit manually on spans, and
highlights the natural structure of the code.

I'm not super attached to the specifics of these commands, so feel free
to suggest changes or even strip them out entirely, but I wanted
something substantial to experiment with when fixing the grafana
integration, and a bug around missing traces on exit.