First pass at tracking system metrics
This spawns a background thread to track common system metrics;
currently only tracks a few, but we can add more as needed.

I evaluated a few different metrics:
 - [opentelemetry-system-metrics](https://crates.io/crates/opentelemetry-system-metrics)
 - [sys_metrics](https://crates.io/crates/sys_metrics)
 - [heim](https://github.com/heim-rs/heim?tab=readme-ov-file)

Of these, sys_metrics doesn't support windows, but is the most actively
maintained.

As a follow up, it might be useful to track Tokio runtime metrics as
well:

https://docs.rs/tokio/latest/tokio/runtime/struct.RuntimeMetrics.html