fix: ensure block counts are reset at the beginning of every epoch.
The blocks count is only truly needed for the reward calculation. It is
only useful at the epoch boundary, and must be reset before the next
epoch starts.

Though it is imperative that this happens AFTER the snapshot has been
taken, since the data pertains (and is required) for the snapshot epoch.