Scaffold chain db import command
Using RocksDB as a backend for the consensus DB means that it's no
longer as simple as a curl to retrieve header data that's needed by
chain_selector to start its process. So having an 'import-chain-db'
command which will connect to a peer and slurp its headers and blocks
seem the next logical step.

Signed-off-by: Arnaud Bailly <arnaud@pankzsoft.com>