Revert "chore: remove now useless with_utxo"
This reverts commit 470432294b0ed2f851ad3ebf8e9550db17022bfd.
Misleading comment from my end during the review; `with_utxo` is
actually used during the import to clear out the database between
imports. This is necessary because importing a new snapshot doesn't
indicate which UTxO have been removed during the epoch, only what the
result is at the end. So we end up keeping many consumed TxO in the
store as a consequence.

Signed-off-by: KtorZ <matthias.benkort@gmail.com>