feat: make header validation possible without crashing
Load a context file that contains information critical for validating
generated headers, and most importantly the starting nonce. This nonce
is injected into the DB at start of simulator as the `active` nonce
for the current epoch, using given header hash which should be the
parent of the first header validated.

Signed-off-by: Arnaud Bailly <arnaud@pankzsoft.com>