fix: fix chain selection for children of genesis (#152)
Due to discrepancies between the representation of parents b/w Rust
and Haskell, the chain selection would not work for blocks immediately
after genesis and therefore would ignore whole chain if it starts from
genesis. By fixing the FakeHeader representation to be more faithful
to actual header, we make the problem surface and can fix it by
handling this edge case.

Signed-off-by: Arnaud Bailly <arnaud@pankzsoft.com>