chore: wire certificates within ledger rules and consume block body during validation.
The idea is to do everything that needs to be done in a single pass, and avoid cloning anything. Elements of the block body are being consumed as they get processed.

This seemingly forces us to group operations together at the moment they occur in the validation. For example, we cannot traverse all certificats once to check their validity, and later, traverse the same list of certificates again to see whether they require some signature witness. Instead, I've augmented the validation context to keep track of this information, notified directly by the appropriate rule. So one can now `require_witness` at any moment during the validation, ensuring that the provided stake credential will be checked for corresponding signatories at the end of the validation.

Signed-off-by: KtorZ <matthias.benkort@gmail.com>