feat: add downstream server
Currently only keepalive and chainsync protocols are supported. There is a smoke test for the latter and the assumption that the former is tested in pallas.
Some work has also been done in the Makefile to make it easier to have multiple Amaru nodes running in the project root directory.
An architecture diagram has been added to the amaru-consensus README.

Signed-Off-By: Roland Kuhn <rk@rkuhn.info>