feat: refactor bootstrap to be testable with in-mem stores
This change makes it possible to test the top-level bootstrap function
that wires all stages together, using in-memory "dummy" stores to
simplify the setup process for ledger and consensus. Those stores are
mostly non-functional but the idea is to make it possible to use those
and enhance their functionalities later in order to ease testing the
whole process using "production" code rather than having to rewire
everything as is currently done in simulator.

Signed-off-by: Arnaud Bailly <arnaud@pankzsoft.com>