fix: validate inputs in lexicographical order
During deserialization, the Haskell node sorts many elements of a
transaction lexicographically. These include the inputs, which are
relevant in this commmit. Pallas intentionally provides only exactly
what's on the wire, which can result in tricky bugs during validation.

While the current solution doesn't feel great, it at least ensures that
the inputs are being processed in the correct order. At some point we
should move the sorting of fields to a different point in the validation
flow, in my opinion.

Signed-off-by: yHSJ <josh@securitybot.info>