feat: TraceBuffer and simulation Replay
Worked on serialization and deserialization of `dyn SendData` and
friends until rustc capitulated. It is now possible to run a simulation
and capture the inputs, states, effects, and their responses in a binary
CBOR log that can later be deserialized again and used to replay a
StageGraph to the same state as the traced simulation run.

Note that currently the log must be complete, otherwise initial states
will be missing and the replay will fail because stage behaviour will be
different.

Note also that while serialization can turn any `dyn SendData` into
byes, it is not possible to get back the same `dyn SendData` by
deserializing a `TraceEntry`. This is impossible because we'd need to
know the actual deserialization code for the concrete underlying type
(which may have type parameters etc.). So the deserialization result is
a generic CBOR `Value` that is then redeserialized in the context where
the concrete value is needed, which is also where the concrete type is
known and the `Deserialize` instance can be captured.

Signed-off-by: Roland Kuhn <rk@rkuhn.info>