feat: set service name at start of daemon
currently, the otel property 'service.name' is hardcoded to 'amaru'
which is confusing when one wants to monitor several amaru services.
This change provides a dead simple solution allowing the process
name to be set at start time and defaulting to 'amaru'. A possibly
more robust solution would use host name resolution but this is not
needed atm.

Signed-off-by: Arnaud Bailly <arnaud@pankzsoft.com>