chore: clarify IncompatibleVersion error message
When trying to connect to a cardano node using an incompatible protocol version
the error message was not explicit:
```
Error: IncompatibleVersion
```

This commit changes that to:
```
failed to connect to peer <peer>: handshake version not accepted
```

It would be even better if we could share the versions supported
by the remote node but we need to make pallas-network share this
information with us so that's probably the best we can do right
now.

Also note that, in the daemon, the first node with which we fail
to connect will make the whole process to stop. Not sure this
is intentional but that's what we had before so I keep it that
way for now.

I'm not super happy with tracing errors and then returning them.
I considered handling errors globally in the `main` function and
I think this is what we should do but, at this point, doing so
reduces the information that we can share and so make diagnosis
a bit difficult. Such an approach would need to define amaru
specific global errors that can be handled at a very high level
but still carry enough information to be usefull... next time ;)

Signed-off-by: Pascal Grange <pascal@grange.nom.fr>