fix: conflate bootstrap witness' roots and verification key hashes
This is weird. Let's talk. Here's a simplified version of the situation:

To verify witnesses, the Haskell node creates two sets:

1. One contains all "public key hashes" collected from looking at addresses.
2. The other contains all "public key hashes" collected from the both the (normal) verification key witnesses, and the bootstrap witnesses.

However, Byron addresses do not contain public key hashes; but a "root", which is a double-hash of some CBOR structure (which somewhere inside, has a public key; but it is impossible to retrieve it from a 'root').

Also, when collecting witnesses, what's being collected in the case of bootstrap witnesses aren't public key hashes, but those re-computed roots.

Now the problem is that this puts in a set (in either case), two objects that have different semantic, and treat them uniformly; all because they are structurally equivalent (28 bytes of data).

Yet, one can now artifically craft a Byron addresses that holds a public key hash instead of a root; and still spend from it by providing a normal verification key witness. From the outside, the transaction appears as missing bootstrap witnesses, but internally, since both witnesses types are conflated, this is approved by the node.

This phenomenon can be witnessed ( :D ! ) on the following PreProd transaction:

- 0c22edee0ffd7c8f32d2fe4da1f144e9ef78dfb51e1678d5198493a83d6cf8ec

which is spending from what looks like Icarus addresses (i.e. gen 2 bootstrap addresses), but doesn't contain any bootstrap witness whatsoever. It does however contains a normal signature witness (even though it's not spending from any standard pub key address). If we hash the key for that witness, we indeed find the root within the address.

Signed-off-by: KtorZ <matthias.benkort@gmail.com>