feat: introduce EpochState trait to abstract populating ledger state
The purpose of this change is to pave the way for populating the
ledger state of Amaru from sources others than CBOR snapshots
extracted from cardano-node, ie. genesis files and data generated for
testnets, Mithril data, or custom snapshots produced by Amaru.

To this end it introduces the concept of an Iterator over
StateFragment that the 'decode_new_epoch_state' function uses to
populate the DB.

Signed-off-by: Arnaud Bailly <arnaud@pankzsoft.com>