feat: can adjust rocksdb settings to match ulimit
When starting Amaru daemon, if file descriptor ulimit is lower
than 65535, Amaru will display a warning to suggest increasing it.

We take the soft limit, we remove 100 to let room for other file
descriptors, we divide by 4 because the daemon usually opens 4
rocksDb instance and allocate that amount of file descriptors
to each rocksdb instance.

Signed-off-by: Pascal Grange <pascal@grange.nom.fr>