chore: Reify a RocksDbConfig
We use several RocksDb instances in the same process. When that's
the case, RocksDb wiki suggests sharing the same `Env` structure
between all the instances. That way, all RocksDb instances share
 a common thread pool, block cache or rate limiter.

To do so, we reify a RocksDbConfig concept. Also, we make it
explicit that we use RocksDb to store data. Before, we had a
`StorePath` that was, either `InMem` or `OnDisk` but now we
we have a `StoreType` that makes it explicit that `OnDisk` is
actually `RocksDb` for two reasons:

1. we may decide to store it on disk but with a different
   technology in the future.
2. more importantly, we need RocksDb specific configuration
   when we use rocksDb so we better with being explicit about
   that.

The shared `Env` is static, initialized with `lazy_static`
which seem fine as it's mainly the daemon that needs it and
it, indeed, keeps RocksDb instances during the whole lifetime
of the process.

See https://github.com/facebook/rocksdb/wiki/RocksDB-Overview#support-for-multiple-embedded-databases-in-the-same-process

Signed-off-by: Pascal Grange <pascal@grange.nom.fr>