feat: initial logic for PlutusData encoding
* wip: TxInfoV1 translation

Signed-off-by: yHSJ <josh@securitybot.info>

* chore: first iteration of constructing TxInfoV1

This iteration uses excessive cloning to create the new struct. In reality, TxInfo (and therefore script contexts) can probably use lifetimes since they will never outlive the transaction itself. That is the next step of the process.

Signed-off-by: yHSJ <josh@securitybot.info>

* wip: ScriptContext type wrappers

We've decided to provide wrappers for types so that we can enforce invariants that are required for the ScriptContext that isn't provided by the underlying types. For example, values must be sorted lexicographically.

Signed-off-by: yHSJ <josh@securitybot.info>

* chore: add mint wrapper

Signed-off-by: yHSJ <josh@securitybot.info>

* chore: add RequiredSigners wrapper

Signed-off-by: yHSJ <josh@securitybot.info>

* chore: add withdrawals wrapper

Signed-off-by: yHSJ <josh@securitybot.info>

* chore: add datums wrapper

Signed-off-by: yHSJ <josh@securitybot.info>

* feat: build a ScriptContext from a TxInfo

Signed-off-by: yHSJ <josh@securitybot.info>

* fix: correctly calculate time range

Signed-off-by: yHSJ <josh@securitybot.info>

---------

Signed-off-by: yHSJ <josh@securitybot.info>