fix: do not compare best_chain_hash and store_chain.roll_forward
Those two values are set at different places in the pipeline and they
might not be necessary in sync: if we ran select_chain several time
before running store_chain, the best_chain_hash will be over the
expected value.

This clearly points to the fact we should only update best chain in
one place, eg. select_chain should not update the DB.

Signed-off-by: Arnaud Bailly <arnaud@pankzsoft.com>