wip: explore creating phase-2 context while performing phase-1 validations
The design is trying to:

  - avoid re-doing work / traverse every structure only once
  - limit cloning of the largest elements (outputs, values, datums, redeemers, scripts)
  - be non-intrusive (built on top of the existing validation context)
  - remain simple to use (piggybacking on Rust's idioms)

So far I've only mapped the inputs, outputs and datums back into a (new definition of) TxInfo. The strategy has been to hook onto the validation context and build the TxInfo while validating; yet, while avoiding to 'pollute' the validation logic.

That allows for minimal changes in the rules, while ensuring that the context now also builds a phase-2 script context along the way.

I am using Arc for the large types with shared ownership. I've tried different approaches but I settled on Rc / Arc ... Anything involving a lifetime is unsuitable because we have nowhere to store the original owned version where references would point to. Ideally, we would store that in the volatile state, but that means the volatile state would have to be self-referencing and that requires some unstable and unsafe rust features. So, to be avoided if possible.

I am a bit less convinced about the last bits. In the sense that we're talking about 16KB of data anyway... so kind of feels overkill here when we could just clone the relevant bits for the script context. But it's also not _too bad_, so maybe worth keeping?

Note also that I got rid of the 'volatile cache' because it was 'getting in the way' and was not really proven useful (maybe even harmful...). We should micro-benchmark this before re-introducing any sort of cache here.

Signed-off-by: KtorZ <matthias.benkort@gmail.com>